all the files are now ready bhai
Only you have to change the background image of front page. Please replace it with any good image. Also, I have updated almost each and every file so one time cross check it or just let it go because i have seen it many a times and found no problems. So, please upload each and every file freshly to hostinger and filezilla because there are changes in almost each and every file of the project DMM.